Implement validation(1/2)

+++ src/Library.Application/Books/Commands/CreateBook/CreateBookCommand.cs
@@ -5,4 +5,4 @@ using MediatR;
 
 namespace Library.Application.Books.Commands.CreateBook;
 
-public record CreateBookCommand(string title, string description, short genreId, Guid authorId, string isbn, int amount, Guid? id = null) : IRequest<ErrorOr<BookDTO>>;
+public record CreateBookCommand(string title, string description, int genreId, Guid authorId, string isbn, int amount, Guid? id = null) : IRequest<ErrorOr<BookDTO>>;
