+++ src/Library.Application/Common/DTO/Books/GenreDTO.cs
@@ -2,6 +2,6 @@ namespace Library.Application.Common.DTO.Books;
 
 public class GenreDTO
 {
-    public short Id { get; private set; }
+    public int Id { get; private set; }
     public string Name { get; private set; }
 }
