+++ src/Library.Application/Genres/Commands/DeleteGenre/DeleteGenreCommand.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace Library.Application.Genres.Commands.DeleteGenre;
 
-public record DeleteGenreCommand(short genreId) : IRequest<ErrorOr<Deleted>>;
+public record DeleteGenreCommand(int genreId) : IRequest<ErrorOr<Deleted>>;
