+++ src/Library.Contracts/Genres/GenreResponse.cs
@@ -1,3 +1,3 @@
 namespace Library.Contracts.Genres;
 
-public record GenreResponse(short id, string name);
+public record GenreResponse(int id, string name);
