+++ src/Library.Domain/Entities/Genre.cs
@@ -2,11 +2,11 @@ namespace Library.Domain.Entities;
 
 public class Genre
 {
-    public short Id { get; private set; }
+    public int Id { get; private set; }
     public string Name { get; private set; }
     public virtual ICollection<Book> Books { get; private set; }
 
-    public Genre(short? id, string name)
+    public Genre(int? id, string name)
     {
         Name = name;
     }
